refactor(updaters): rename repositories/updaters folder
move reducer*.js in new updater/ folder

Ref #4
